Restructure to fix data checking between meals
This is basically recording the name of the meal when upgrading to main

This state is passed to a random meal being generated.

This is to determine that both are indeed different.

It's a bit trickier in the beginning when both meals are generated

The button which handles the transition to main is also relevant

It stores the meal name as a prop so that if chosen as main..

..then that would be the name to be checked against.